refactor: clearly identify data added by dinghy
The JSON data rendered by templates is mostly the GraphQL data returned
by GitHub.  But there is a bit of computed data added by dinghy, and
some pieces have been re-parented to indicate hierarchy.  All of the
keys added by dinghy now have a `dinghy_` prefix for clarity.

This is in support of issues #31 and #32.